requirements.txt
flet~=0.23.2
pandas~=2.2.2
python-dotenv~=1.0.1
XlsxWriter~=3.2.0
bs4~=0.0.2
beautifulsoup4~=4.12.3
selenium~=4.23.1
cx_Freeze~=7.2.0
pyinstaller~=6.10.0
winshell~=0.6
pypdf2~=3.0.1
# PyPDF2~=3.0.1